Move `TransferMessageStorage`
